refactor: add getData method for experiment and component fetch
