+++ ios/Nativebrik/Nativebrik/data/component.swift
@@ -16,28 +16,23 @@ class ComponentRepositoryImpl: ComponentRepository2 {
     init(config: Config) {
         self.config = config
     }
-    
+
     func fetchComponent(experimentId: String, id: String) async -> Result<UIBlock, NativebrikError> {
         guard let url = URL(string: config.cdnUrl + "/projects/" + config.projectId + "/experiments/components/" + experimentId + "/" + id) else {
             return Result.failure(NativebrikError.irregular("Failed to create URL object"))
         }
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        do {
-            let (data, response) = try await nativebrikSession.data(for: request)
-            guard let res = response as? HTTPURLResponse else {
-                return Result.failure(NativebrikError.irregular("Failed to parse as HttpURLResponse"))
-            }
-            if res.statusCode == 404 {
-                return Result.failure(NativebrikError.notFound)
-            }
+        
+        let data = await getData(url: url)
+        switch data {
+        case .success(let data):
             let decoder = JSONDecoder()
             guard let result = try? decoder.decode(UIBlockJSON.self, from: data) else {
                 return Result.failure(NativebrikError.failedToDecode)
             }
             return Result.success(result)
-        } catch {
-            return Result.failure(NativebrikError.other(error))
+            
+        case .failure(let error):
+            return Result.failure(error)
         }
     }
 }
