+++ ios/Nativebrik/Nativebrik/data/experiment.swift
@@ -22,25 +22,18 @@ class ExperimentRepositoryImpl: ExperimentRepository2 {
         guard let url = URL(string: self.config.cdnUrl + "/projects/" + self.config.projectId + "/experiments/id/" + id) else {
             return Result.failure(NativebrikError.irregular("Failed to create URL object"))
         }
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        do {
-            let t0 = getCurrentDate()
-            let (data, response) = try await nativebrikSession.data(for: request)
-            guard let res = response as? HTTPURLResponse else {
-                return Result.failure(NativebrikError.irregular("Failed to parse as HttpURLResponse"))
-            }
-            syncDateFromHTTPURLResponse(t0: t0, res: res)
-            if res.statusCode == 404 {
-                return Result.failure(NativebrikError.notFound)
-            }
+        
+        let data = await getData(url: url, syncDateTime: true)
+        switch data {
+        case .success(let data):
             let decoder = JSONDecoder()
             guard let result = try? decoder.decode(ExperimentConfigs.self, from: data) else {
                 return Result.failure(NativebrikError.failedToDecode)
             }
             return Result.success(result)
-        } catch {
-            return Result.failure(NativebrikError.other(error))
+            
+        case .failure(let error):
+            return Result.failure(error)
         }
     }
     
@@ -48,25 +41,18 @@ class ExperimentRepositoryImpl: ExperimentRepository2 {
         guard let url = URL(string: config.cdnUrl + "/projects/" + config.projectId + "/experiments/trigger/" + name) else {
             return Result.failure(NativebrikError.irregular("Failed to create URL object"))
         }
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        do {
-            let t0 = getCurrentDate()
-            let (data, response) = try await nativebrikSession.data(for: request)
-            guard let res = response as? HTTPURLResponse else {
-                return Result.failure(NativebrikError.irregular("Failed to parse as HttpURLResponse"))
-            }
-            syncDateFromHTTPURLResponse(t0: t0, res: res)
-            if res.statusCode == 404 {
-                return Result.failure(NativebrikError.notFound)
-            }
+        
+        let data = await getData(url: url, syncDateTime: true)
+        switch data {
+        case .success(let data):
             let decoder = JSONDecoder()
             guard let result = try? decoder.decode(ExperimentConfigs.self, from: data) else {
                 return Result.failure(NativebrikError.failedToDecode)
             }
             return Result.success(result)
-        } catch {
-            return Result.failure(NativebrikError.other(error))
+            
+        case .failure(let error):
+            return Result.failure(error)
         }
     }
 }
